requirements.txt
flask
tensorflow-cpu
opencv-python
numpy
gunicorn